Rename view files grant_material_association -> grant_material
