+++ api/resources_portal/views/__init__.py
@@ -4,7 +4,7 @@ from resources_portal.views.document_views import (
     UserDocumentView,
 )
 from resources_portal.views.grant import GrantViewSet
-from resources_portal.views.grant_material_association import GrantMaterialViewSet
+from resources_portal.views.grant_material import GrantMaterialViewSet
 from resources_portal.views.organization_member import OrganizationMemberViewSet
 from resources_portal.views.material import MaterialViewSet
 from resources_portal.views.organization import OrganizationViewSet
