+++ api/resources_portal/views/grant_material.py
